rename viem_public
